Moved test files into separate directories
