+++ package.json
@@ -5,7 +5,7 @@
   "main": "src/index.js",
   "scripts": {
     "start": "node src/index.js",
-    "test": "mocha test/"
+    "test": "mocha test/ --recursive"
   },
   "repository": {
     "type": "git",
