+++ test/10_database/databaseManager.js
@@ -1,11 +1,11 @@
-import {DataBaseManager} from "../src/database/DatabaseManager.js"
+import {DataBaseManager} from "../../src/database/DatabaseManager.js"
 import chai from 'chai';
-import {Administrator} from "../src/database/modele/Users.js";
-import {Discussion} from "../src/database/modele/Discussion.js";
-import {Question} from "../src/database/modele/Question.js";
-import {Device} from "../src/database/modele/Device.js";
-import {Response} from "../src/database/modele/Response.js";
-import {logger} from "../src/conf/config.js";
+import {Administrator} from "../../src/database/modele/Users.js";
+import {Discussion} from "../../src/database/modele/Discussion.js";
+import {Question} from "../../src/database/modele/Question.js";
+import {Device} from "../../src/database/modele/Device.js";
+import {Response} from "../../src/database/modele/Response.js";
+import {logger} from "../../src/conf/config.js";
 
 const expect = chai.expect;
 const should = chai.should();
