--- test/sample-test.js
@@ -1,8 +1,0 @@
-import chai from 'chai';
-
-let expect = chai.expect;
-
-it('Sample always valid test', () => {
-    let foo = true;
-    expect(foo).to.equal(true);
-});
